backend/requirements.txt

fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
openai==1.3.7
google-generativeai==0.3.2
chromadb==0.4.18
PyMuPDF==1.23.8
requests==2.31.0
pydantic==2.5.0
pydantic-settings==2.1.0
